img .png -> .web and move img preview to comp
